feat : Contract 상세 API 응답 클래스 변수 타입 변경, 추가

+++ src/main/java/kr/co/houmuch/api/domain/dto/contract/ContractSummary.java
@@ -10,8 +10,8 @@ import lombok.experimental.SuperBuilder;
 @AllArgsConstructor
 @SuperBuilder
 public abstract class ContractSummary {
-    protected int priceMonth;
-    protected int priceYear;
+    protected double priceMonth;
+    protected double priceYear;
     protected int tradeCount;
     protected int rentCount;
 }
